+++ packages/playground-ui/src/domains/observability/components/index.ts
@@ -1,6 +1,6 @@
-export * from './trace-tree';
-export * from './trace-tree-legend';
-export * from './trace-tree-span';
+export * from './trace-timeline';
+export * from './trace-timeline-legend';
+export * from './trace-timeline-span';
 export * from './shared';
 export * from './trace-dialog';
 export * from './traces-tools';
